@ refactor(866): carve SuspendNonCompliantMembersJob body into Humans.Users (G5 lane 4b-2d, part 2)
The step-6b carve, forced by finding 31. SuspendNonCompliantMembersJob is
a serialized Hangfire target — RecurringJobExtensions registers it as
RecurringJob.AddOrUpdate<SuspendNonCompliantMembersJob>, which pins the
declaring type name to its assembly and namespace. A job queued or
retry-delayed across a deploy that moved it could not resolve its target
afterwards, and neither the build nor the suite would say so.

So the type does not move. Its assembly-qualified name is unchanged:
Humans.Infrastructure.Jobs.SuspendNonCompliantMembersJob, Humans.Infrastructure.
What moves is the body, behind INonCompliantMemberSuspension on
Humans.Users.Contracts; the implementation is resolved from DI at execution
time, so this is safe by construction. What is left in Base is the start
log, a try/catch and the failure metric — its ten cross-section contract
injections are gone.

The contract is "do the thing", not "give me the rows". Every metric call
and log line fires exactly when it did before, including the pre-existing
asymmetry where the no-candidates early return records no job-run metric;
that is preserved rather than tidied.

nameof(SuspendNonCompliantMembersJob) was the persisted audit actor. The
section cannot name a Humans.Infrastructure type, so it becomes a pinned
literal beside a comment (memory/code/type-name-as-persisted-string.md) —
the rows already in the database keep matching.

Humans.Users gains two project references, Humans.Governance.Contracts and
Humans.GoogleIntegration.Contracts. Both leaves reference only
Humans.Interfaces and Humans.Domain, so both edges are acyclic.

The job test becomes NonCompliantMemberSuspensionTests in the section.
reforge.surface-score.json (live config, consumed by pr-surface-report.yml)
and four docs carrying the old paths are repointed.

No schema change, no migration, no route change.

Refs nobodies-collective/Humans#866

@
