+++ src/Humans.Infrastructure/Jobs/SuspendNonCompliantMembersJob.cs
@@ -1,18 +1,7 @@
-using Humans.GoogleIntegration.Contracts;
 using Hangfire;
 using Microsoft.Extensions.Logging;
 using NodaTime;
 using Humans.Application.Interfaces;
-using Humans.AuditLog.Contracts;
-using Humans.Application.Interfaces.Caching;
-using Humans.Email.Contracts;
-using Humans.Governance.Contracts;
-using Humans.Notifications.Contracts;
-using Humans.Shifts.Contracts;
-using Humans.Teams.Contracts;
-using Humans.Application.Interfaces.Users;
-using Humans.Domain.Entities;
-using Humans.Domain.Enums;
 using Humans.Users.Contracts;
 
 namespace Humans.Infrastructure.Jobs;
@@ -22,28 +11,22 @@ namespace Humans.Infrastructure.Jobs;
 /// after the grace period has expired.
 /// </summary>
 /// <remarks>
-/// All reads/writes fan out through section services
-/// (<see cref="IUserService"/>,
-/// <see cref="ITeamService"/>, <see cref="IGoogleSyncService"/>) so the job
-/// never touches a section DbContext
-/// directly (design-rules §2c). Cross-cutting cache invalidation routes
-/// through invalidator interfaces
-/// (<see cref="IRoleAssignmentClaimsCacheInvalidator"/>,
-/// <see cref="IShiftAuthorizationInvalidator"/>) rather than IMemoryCache.
+/// A shim. The body moved into <c>Humans.Users</c> at G5 lane 4b-2d (Peter, 2026-08-14:
+/// membership lifecycle is Users, not Governance) and is reached through
+/// <see cref="INonCompliantMemberSuspension"/>.
+///
+/// <b>This type must not move, be renamed, or change namespace.</b> Hangfire serializes the
+/// declaring type name of a recurring job target, and
+/// <c>RecurringJob.AddOrUpdate&lt;SuspendNonCompliantMembersJob&gt;</c> in
+/// <c>RecurringJobExtensions</c> pins it to this assembly and namespace. A job enqueued or
+/// retry-delayed before a deploy that relocated it would fail to resolve its target
+/// afterwards, and neither the build nor the test suite would report it
+/// (G5 batch #4 finding 31). Carving the body out is safe by construction: the serialized
+/// type never moves and the implementation is resolved from DI at execution time.
 /// </remarks>
 [DisableConcurrentExecution(timeoutInSeconds: 300)]
 public class SuspendNonCompliantMembersJob(
-    IUserService userService,
-    ITeamServiceRead teamService,
-    IActiveTeamsCacheInvalidator activeTeamsCacheInvalidator,
-    IMembershipCalculatorRead membershipCalculator,
-    IEmailService emailService,
-    IEmailMessageFactory emailMessages,
-    INotificationEmitter notificationService,
-    IGoogleSyncService googleSyncService,
-    IAuditLogService auditLogService,
-    IRoleAssignmentClaimsCacheInvalidator roleAssignmentClaimsInvalidator,
-    IShiftAuthorizationInvalidator shiftAuthorizationInvalidator,
+    INonCompliantMemberSuspension suspension,
     IHumansMetrics metrics,
     ILogger<SuspendNonCompliantMembersJob> logger,
     IClock clock) : IRecurringJob
@@ -59,127 +42,7 @@ public class SuspendNonCompliantMembersJob(
 
         try
         {
-            // Get users who are now Inactive (missing consents + grace period expired)
-            var usersToSuspend = await membershipCalculator
-                .GetUsersRequiringStatusUpdateAsync(cancellationToken);
-
-            if (usersToSuspend.Count == 0)
-            {
-                logger.LogInformation("Completed suspension check, no users require suspension");
-                return;
-            }
-
-            var now = clock.GetCurrentInstant();
-
-            // Apply the suspension write through IUserService — returns the
-            // subset of user ids whose profile was actually mutated (skips
-            // already-suspended / profileless users).
-            var suspendedIds = await userService
-                .SuspendProfilesForMissingConsentAsync(usersToSuspend, now, cancellationToken);
-
-            if (suspendedIds.Count == 0)
-            {
-                metrics.RecordJobRun("suspend_noncompliant_members", "success");
-                logger.LogInformation(
-                    "Completed non-compliant member check, no eligible users to suspend");
-                return;
-            }
-
-            // Fan out user + email hydration for notifications, and team membership
-            // lookup for Google-sync cleanup.
-            var usersById = await userService
-                .GetUserInfosAsync(suspendedIds, cancellationToken);
-
-            foreach (var userId in suspendedIds)
-            {
-                if (!usersById.TryGetValue(userId, out var user))
-                {
-                    logger.LogWarning(
-                        "Suspended user {UserId} not found in user lookup — skipping downstream side effects",
-                        userId);
-                    continue;
-                }
-
-                // 1. Send email notification
-                var effectiveEmail = user.Email;
-                if (effectiveEmail is not null)
-                {
-                    try
-                    {
-                        await emailService.SendAsync(emailMessages.AccessSuspended(
-                            effectiveEmail,
-                            user.BurnerName,
-                            "Missing required document consent (grace period expired)",
-                            user.PreferredLanguage),
-                            cancellationToken);
-                    }
-                    catch (Exception ex)
-                    {
-                        logger.LogError(ex, "Failed to send suspension email for user {UserId}", user.Id);
-                    }
-                }
-
-                // 2. Send in-app notification (best-effort)
-                try
-                {
-                    await notificationService.SendAsync(
-                        NotificationSource.AccessSuspended,
-                        NotificationClass.Actionable,
-                        NotificationPriority.Critical,
-                        "Your access has been suspended",
-                        [user.Id],
-                        body: "Your access has been suspended because required document consent is missing. Please review and sign the required documents to restore access.",
-                        actionUrl: "/Consent",
-                        actionLabel: "Review documents",
-                        cancellationToken: cancellationToken);
-                }
-                catch (Exception ex)
-                {
-                    logger.LogError(ex, "Failed to dispatch AccessSuspended notification for user {UserId}", user.Id);
-                }
-
-                // 3. Remove from all team resources (Google Drive/Groups) for the user's active teams.
-                var memberTeamIds = (await teamService.GetTeamsAsync(cancellationToken)).Values
-                    .Where(t => t.Members.Any(m => m.UserId == user.Id))
-                    .Select(t => t.Id)
-                    .ToList();
-                foreach (var teamId in memberTeamIds)
-                {
-                    try
-                    {
-                        await googleSyncService.RemoveUserFromTeamResourcesAsync(
-                            teamId,
-                            user.Id,
-                            cancellationToken);
-                    }
-                    catch (Exception ex)
-                    {
-                        logger.LogError(ex, "Failed to remove user {UserId} from team {TeamId} resources during suspension",
-                            user.Id, teamId);
-                    }
-                }
-
-                logger.LogWarning(
-                    "User {UserId} ({Email}) suspended and flagged for removal from {Count} teams",
-                    user.Id, effectiveEmail, memberTeamIds.Count);
-
-                metrics.RecordMemberSuspended("job");
-
-                // 4. Audit log + cross-cutting cache invalidation.
-                await auditLogService.LogAsync(
-                    AuditAction.MemberSuspended, nameof(User), user.Id,
-                    $"{user.BurnerName} suspended for missing required document consent (grace period expired)",
-                    nameof(SuspendNonCompliantMembersJob));
-
-                roleAssignmentClaimsInvalidator.Invalidate(user.Id);
-                shiftAuthorizationInvalidator.Invalidate(user.Id);
-                activeTeamsCacheInvalidator.Invalidate();
-            }
-
-            metrics.RecordJobRun("suspend_noncompliant_members", "success");
-            logger.LogInformation(
-                "Completed non-compliant member check, suspended {Count} members",
-                suspendedIds.Count);
+            await suspension.SuspendNonCompliantAsync(cancellationToken);
         }
         catch (Exception ex)
         {
